LayoutTests/svg/dynamic-updates/SVGFESpotLightElement-dom-x-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spotLightElement.getAttribute('x') is "0"
PASS spotLightElement.getAttribute('x') is "113"
PASS successfullyParsed is true

TEST COMPLETE
